IExecutorService cleanup
