+++ hazelcast/src/main/java/com/hazelcast/executor/impl/client/TargetCallableRequest.java
@@ -19,7 +19,7 @@ package com.hazelcast.executor.impl.client;
 import com.hazelcast.client.impl.client.TargetClientRequest;
 import com.hazelcast.executor.impl.DistributedExecutorService;
 import com.hazelcast.executor.impl.ExecutorPortableHook;
-import com.hazelcast.executor.impl.MemberCallableTaskOperation;
+import com.hazelcast.executor.impl.operations.MemberCallableTaskOperation;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
@@ -29,6 +29,7 @@ import com.hazelcast.nio.serialization.PortableWriter;
 import com.hazelcast.security.SecurityContext;
 import com.hazelcast.spi.Operation;
 
+import javax.security.auth.Subject;
 import java.io.IOException;
 import java.security.Permission;
 import java.util.concurrent.Callable;
@@ -58,7 +59,8 @@ public final class TargetCallableRequest extends TargetClientRequest {
     protected Operation prepareOperation() {
         SecurityContext securityContext = getClientEngine().getSecurityContext();
         if (securityContext != null) {
-            callable = securityContext.createSecureCallable(getEndpoint().getSubject(), callable);
+            Subject subject = getEndpoint().getSubject();
+            callable = securityContext.createSecureCallable(subject, callable);
         }
         Data callableData = serializationService.toData(callable);
         return new MemberCallableTaskOperation(name, uuid, callableData);
